requirements.txt
openai
ollama
gradio
groq
pydub
python-dotenv
youtube_transcript_api